pyproject.toml
[tool.poetry]
name = "bachelorproject"
version = "0.1.0"
description = ""
authors = ["Matthijs <matthijs.vanderlende@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pandas = "^2.1.3"
more-itertools = "^10.1.0"
scipy = "^1.12.0"
gpytorch="^1.11"
matplotlib="^3.8.2"
notebook="^7.1.0"
ipykernel="^6.29.2"
loguru = "^0.7.2"
skrl = "^1.1.0"
tensorboard = "^2.16.2"
stable-baselines3 = "^2.2.1"
ale-py = "^0.8.1"
gymnasium = {version = "^0.29.1", extras = ["box2d", "atari", "accept-rom-license"]}
opencv-python = "^4.9.0.80"
moviepy = "^1.0.3"
pyav = "^12.0.5"
torch = "^2.3.0"
torchvision = "^0.18.0"
torchrl = "^0.4.0"
hydra-core = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
